[skip uplift] Add custom eltwise binary file to be used in SDPA optimizations (#1269)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/37044
### Problem description
Minimize ADDR_MODs overlap between ops

### What's changed


### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Refactoring

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)
- [ ] [Assert validation](docs/Introduction_to_asserts.md) Complied with
assert doc (if applicable)

---------

+++ tests/sources/matmul_custom_test.cpp
@@ -72,14 +72,14 @@ void run_kernel(const volatile struct RuntimeParams *params)
 
 void run_kernel(const volatile struct RuntimeParams *params)
 {
-    _llk_math_matmul_init_no_mop_<static_cast<int>(MATH_FIDELITY)>(
+    _llk_math_matmul_init_no_mop_<MATH_FIDELITY>(
         TILE_R_DIM, TILE_C_DIM, TILE_R_DIM, TILE_C_DIM, false /* partial_face */, 0 /* transpose */, params->CT_DIM, params->RT_DIM);
     _llk_math_pack_sync_init_<DstSync::SyncHalf, is_fp32_dest_acc_en>();
     _llk_math_hw_configure_<is_fp32_dest_acc_en>(formats.math, formats.math);
     _llk_math_wait_for_dest_available_<DstSync::SyncHalf>();
     for (std::uint32_t j = 0; j < params->KT_DIM; j++)
     {
-        _llk_math_matmul_no_mop_<static_cast<int>(MATH_FIDELITY)>(0 /* dst_index */, params->CT_DIM, params->RT_DIM);
+        _llk_math_matmul_no_mop_<MATH_FIDELITY>(0 /* dst_index */, params->CT_DIM, params->RT_DIM);
     }
     _llk_math_dest_section_done_<DstSync::SyncHalf, is_fp32_dest_acc_en>();
 }
